[i] remocao de parte das dependencias para o pysus [i] alteracao em processo de coleta para criar pastas considerando estado, ano, mes, tipo de arquivo e grupo
