+++ auto-get-files/collect_from_ftp.sh
